Add camera icon
Icons are included via a partial (see `/_includes/_icon.html`. Their
entries in `/_icons` are actually markdown files with an .svg
extension, with YAML for the alt text / title,  and the content is just
the path(s) required for the vector art. The viewbox and all other svg
info is controlled in the partial, for consistency and to keep the svgs
interactive (manipulatable via CSS and JS) and lightweight.

The current solution doesn’t yet have a PNG fallback, its on the list
to investigate if needed for our browser support.