Update to a department's site link & the format of hours
The original language indicated that the department’s page would be
hosted on austintexas.gov, but some departments (such as Austin Energy)
have their own separate domains. The language has been updated to be a
more general “Learn more about [department name]”. Also added in a more
descriptive example for how we want the hours to be written out to be
more accessible.