Update header navigation styling
I removed the blue link styling from the header links so they didn’t
look so prominent. I updated the “View…” link at the bottom of the
lists to align with the footer “View…” visual style. I also updated the
background color of the header nav to white to give the whole header
area one color—it was looking very stripe-y. The rollover state is now
a more subtle gray-lightest.