Add a separate dev config file and remove plugins to speed up site generation
Really, we don’t need the jekyll-sitemap plugin working locally and it
takes 20s or so to build, and the plugin maintainers don’t give a way
to disable it. [Their recommendation of excluding the
gem](https://github.com/jekyll/jekyll-sitemap/issues/60) requires
actually uninstalling it so that it’s not in the Gemfile, a change
which can’t be committed back into the repo and still have the sitemap
work on production. Still, having separate config files does keep some
things separate, like the algolia info, which is only pushed locally
right now.
To use the dev config file instead:

`bundle exec jekyll serve --config "_config_dev.yml" --profile`

I’ll probably make a local change to remove the Jekyll-Sitemap gem in
the gem file and never commit it, so that it doesn’t take forever to
build :)