Move all pages to be within /residential folder
At first I tried this with `site.baseurl`, but that moved all the
assets and everything (dozens of pdfs, not to mention all the
images/css/js) and didn’t give me a clean, jekyll-y way to do a
redirect from the index page. I also removed the absolute urls wherever
they weren’t needed, so that it’d be easier to change to baseurl later
if we want to share this domain with something else, or use this same
Jekyll install to do /commercial permitting pages.