Add files from CloudCannon
It seems like syncing with CloudCannon has been broken for awhile. The suggested fix is to reconnect the github repo, but doing so means that CloudCannon will sync the files from GitHub first. That means we'll get a bunch of outdated crap.

This is a backup of the latest changes from CloudCannon. So reconnecting the repo will just pull in this stuff. I hope.